Building sandmark's packages with "opam install"
Delete `Sand_build because it the same with `Opam.
Build sandmark's packages by using "opam install" after skipping the analysis.

During the build of a package we pin all packages in the same repo.
Important when code is from a cloned repo because we could end up
building from opam source.